# docs/system/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseIdResultDTO": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdResultDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseObject": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseStockPageDTO": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StockPageDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "IdResultDTO": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ItemCreateDTO": {
        "properties": {
          "baseUnit": {
            "maxLength": 32,
            "minLength": 0,
            "type": "string"
          },
          "code": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "baseUnit",
          "code",
          "name"
        ],
        "type": "object"
      },
...
      "LocationUpdateDTO": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "version"
        ],
        "type": "object"
      },
...
      "StockPageDTO": {
        "properties": {
          "current": {
            "format": "int64",
            "type": "integer"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/StockPageItemDTO"
            },
            "type": "array"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "StockPageItemDTO": {
        "properties": {
          "baseUnit": {
            "type": "string"
          },
          "itemCode": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "locationCode": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "locationName": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          },
          "warehouseCode": {
            "type": "string"
          },
          "warehouseId": {
            "type": "string"
          },
          "warehouseName": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "WarehouseCreateDTO": {
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "departmentId": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "code",
          "departmentId",
          "name"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "由 contract profile 在运行时生成的 0.1 API 契约。",
    "title": "Internal Admin Template API",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/warehouse/items": {
      "get": {
        "operationId": "items",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      },
      "post": {
        "operationId": "createItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIdResultDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
...
    "/api/warehouse/locations/{id}": {
      "put": {
        "operationId": "updateLocation",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
    "/api/warehouse/locations/{warehouseId}": {
      "get": {
        "operationId": "locations",
        "parameters": [
          {
            "in": "path",
            "name": "warehouseId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
    "/api/warehouse/movements/recent": {
      "get": {
        "operationId": "movements",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
    "/api/warehouse/operations": {
      "get": {
        "operationId": "operations",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
    "/api/warehouse/operations/{operationId}": {
      "get": {
        "operationId": "operation",
        "parameters": [
          {
            "in": "path",
            "name": "operationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
    "/api/warehouse/operations/{operationId}/movements": {
      "get": {
        "operationId": "operationMovements",
        "parameters": [
          {
            "in": "path",
            "name": "operationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
...
    "/api/warehouse/stock": {
      "get": {
        "operationId": "stockPage",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "itemId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locationId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStockPageDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
    "/api/warehouse/stock/items/{itemId}": {
      "get": {
        "operationId": "stock",
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
    "/api/warehouse/stock/locations/{locationId}": {
      "get": {
        "operationId": "contents",
        "parameters": [
          {
            "in": "path",
            "name": "locationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      }
    },
...
    "/api/warehouse/warehouses": {
      "get": {
        "operationId": "warehouses",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-query-controller"
        ]
      },
      "post": {
        "operationId": "createWarehouse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIdResultDTO"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "warehouse-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}